requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==7.1.2
Flask==1.1.2
frozenlist==1.3.0
gunicorn==20.1.0
idna==3.3
ipinfo==4.2.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==6.0.2
psycopg2==2.9.1
python-dotenv==0.19.2
PyYAML==6.0
requests==2.27.1
typing-extensions==4.0.1
ua-parser==0.10.0
urllib3==1.26.8
user-agents==2.2.0
Werkzeug==1.0.1
yarl==1.7.2
